+++ src/TimeScheduler/System/Threading/TimeProviderCancellationTokenSourceExtensions.cs
@@ -1,5 +1,4 @@
 using System.Runtime.CompilerServices;
-using TimeScheduler;
 
 namespace System.Threading;