+++ test/TimeScheduler.Tests/System/Testing/ManualTimeProviderDelayTests.cs
@@ -1,13 +1,13 @@
 namespace System.Testing;
 
-public class TestTimeProviderDelayTests
+public class ManualTimeProviderDelayTests
 {
     [Fact]
     public void Delayed_task_is_completes()
     {
         var startTime = DateTimeOffset.UtcNow;
         var future = TimeSpan.FromTicks(1);
-        using var sut = new TestTimeProvider(startTime);
+        var sut = new ManualTimeProvider(startTime);
         var task = sut.Delay(TimeSpan.FromTicks(1));
 
         sut.ForwardTime(future);
@@ -19,7 +19,7 @@ public class TestTimeProviderDelayTests
     public void Delayed_task_is_cancelled()
     {
         using var cts = new CancellationTokenSource();
-        using var sut = new TestTimeProvider();
+        var sut = new ManualTimeProvider();
         var task = sut.Delay(TimeSpan.FromTicks(1), cts.Token);
 
         cts.Cancel();