+++ test/TimeScheduler.Tests/System/Testing/ManualTimeProviderPeriodicTimerTests.cs
@@ -1,14 +1,12 @@
-using System.Threading.Tasks;
-
 namespace System.Testing;
 
-public class TestTimeProviderPeriodicTimerTests
+public class ManualTimeProviderPeriodicTimerTests
 {
     [Fact]
     public void PeriodicTimer_WaitForNextTickAsync_cancelled_immediately()
     {
         using var cts = new CancellationTokenSource();
-        using var sut = new TestTimeProvider();
+        var sut = new ManualTimeProvider();
         using var periodicTimer = sut.CreatePeriodicTimer(TimeSpan.FromTicks(1));
 
         cts.Cancel();
@@ -20,7 +18,7 @@ public class TestTimeProviderPeriodicTimerTests
     [Fact]
     public async Task PeriodicTimer_WaitForNextTickAsync_complete_immediately()
     {
-        using var sut = new TestTimeProvider();
+        var sut = new ManualTimeProvider();
         using var periodicTimer = sut.CreatePeriodicTimer(TimeSpan.FromTicks(1));
 
         sut.ForwardTime(TimeSpan.FromTicks(1));
@@ -34,7 +32,7 @@ public class TestTimeProviderPeriodicTimerTests
     {
         var startTime = DateTimeOffset.UtcNow;
         var future = TimeSpan.FromTicks(1);
-        using var sut = new TestTimeProvider(startTime);
+        var sut = new ManualTimeProvider(startTime);
         using var periodicTimer = sut.CreatePeriodicTimer(TimeSpan.FromTicks(1));
         var task = periodicTimer.WaitForNextTickAsync();
 
@@ -47,7 +45,7 @@ public class TestTimeProviderPeriodicTimerTests
     public async Task PeriodicTimer_WaitForNextTickAsync_completes_after_dispose()
     {
         var startTime = DateTimeOffset.UtcNow;
-        using var sut = new TestTimeProvider(startTime);
+        var sut = new ManualTimeProvider(startTime);
         var periodicTimer = sut.CreatePeriodicTimer(TimeSpan.FromTicks(1));
         var task = periodicTimer.WaitForNextTickAsync();
 
@@ -60,7 +58,7 @@ public class TestTimeProviderPeriodicTimerTests
     public async Task PeriodicTimer_WaitForNextTickAsync_cancelled_with_exception()
     {
         using var cts = new CancellationTokenSource();
-        using var sut = new TestTimeProvider();
+        var sut = new ManualTimeProvider();
         using var periodicTimer = sut.CreatePeriodicTimer(TimeSpan.FromTicks(1));
         var task = periodicTimer.WaitForNextTickAsync(cts.Token);
         cts.CancelAfter(TimeSpan.Zero);
@@ -75,7 +73,7 @@ public class TestTimeProviderPeriodicTimerTests
     [Fact]
     public void PeriodicTimer_WaitForNextTickAsync_twice_throws()
     {
-        using var sut = new TestTimeProvider();
+        var sut = new ManualTimeProvider();
         using var periodicTimer = sut.CreatePeriodicTimer(TimeSpan.FromTicks(1));
 
         _ = periodicTimer.WaitForNextTickAsync();
@@ -87,30 +85,34 @@ public class TestTimeProviderPeriodicTimerTests
     [Fact]
     public void PeriodicTimer_WaitForNextTickAsync_completes_multiple()
     {
-        using var sut = new TestTimeProvider();
+        var sut = new ManualTimeProvider();
         var calledTimes = 0;
-        var looper = WaitForNextTickInLoop(sut, () => calledTimes++);
+        var interval = TimeSpan.FromSeconds(1);
+        var looper = WaitForNextTickInLoop(sut, () => calledTimes++, interval);
 
-        sut.ForwardTime(TimeSpan.FromSeconds(1));
+        sut.ForwardTime(interval);
         calledTimes.Should().Be(1);
 
-        sut.ForwardTime(TimeSpan.FromSeconds(1));
+        sut.ForwardTime(interval);
         calledTimes.Should().Be(2);
+    }
 
-        static async Task WaitForNextTickInLoop(TimeProvider scheduler, Action callback)
-        {
-            using var periodicTimer = scheduler.CreatePeriodicTimer(TimeSpan.FromSeconds(1));
-            while (await periodicTimer.WaitForNextTickAsync(CancellationToken.None))
-            {
-                callback();
-            }
-        }
+    [Fact]
+    public void PeriodicTimer_WaitForNextTickAsync_completes_multiple_single_forward()
+    {
+        var sut = new ManualTimeProvider();
+        var calledTimes = 0;
+        var interval = TimeSpan.FromSeconds(1);
+        var looper = WaitForNextTickInLoop(sut, () => calledTimes++, interval);
+
+        sut.ForwardTime(interval * 3);
+        calledTimes.Should().Be(3);
     }
 
     [Fact]
     public async void PeriodicTimer_WaitForNextTickAsync_exists_on_timer_Dispose()
     {
-        using var sut = new TestTimeProvider();
+        var sut = new ManualTimeProvider();
         var periodicTimer = sut.CreatePeriodicTimer(TimeSpan.FromSeconds(1));
         var disposeTask = WaitForNextTickToReturnFalse(periodicTimer);
         sut.ForwardTime(TimeSpan.FromSeconds(1));
@@ -126,6 +128,32 @@ public class TestTimeProviderPeriodicTimerTests
             }
 
             return false;
+        }
+    }
+
+    [Fact]
+    public void GetUtcNow_matches_time_when_WaitForNextTickAsync_is_invoked()
+    {
+        var sut = new ManualTimeProvider();
+        var startTime = sut.GetUtcNow();
+        var callbackTimes = new List<DateTimeOffset>();
+        var interval = TimeSpan.FromSeconds(5);
+        var looper = WaitForNextTickInLoop(sut, () => callbackTimes.Add(sut.GetUtcNow()), interval);
+
+        sut.ForwardTime(interval * 3);
+
+        callbackTimes.Should().Equal(
+            startTime + interval * 1,
+            startTime + interval * 2,
+            startTime + interval * 3);
+    }
+
+    static async Task WaitForNextTickInLoop(TimeProvider scheduler, Action callback, TimeSpan interval)
+    {
+        using var periodicTimer = scheduler.CreatePeriodicTimer(interval);
+        while (await periodicTimer.WaitForNextTickAsync(CancellationToken.None))
+        {
+            callback();
         }
     }
 }