+++ test/TimeScheduler.Tests/Testing/TestSchedulerTimerTests.cs
@@ -8,7 +8,7 @@ public class TestSchedulerTimerTests
         var callbackCount = 0;
         var dueTime = TimeSpan.FromSeconds(1);
         var period = Timeout.InfiniteTimeSpan;
-        using var sut = new TestScheduler();
+        var sut = new TestScheduler();
         using var timer = sut.CreateTimer(_ => callbackCount++, null, dueTime, period);
 
         sut.ForwardTime(dueTime);
@@ -24,7 +24,7 @@ public class TestSchedulerTimerTests
         var callbackCount = 0;
         var dueTime = TimeSpan.FromSeconds(1);
         var period = TimeSpan.FromSeconds(2);
-        using var sut = new TestScheduler();
+        var sut = new TestScheduler();
         using var timer = sut.CreateTimer(_ => callbackCount++, null, dueTime, period);
 
         sut.ForwardTime(dueTime);
@@ -43,7 +43,7 @@ public class TestSchedulerTimerTests
         var callbackCount = 0;
         var dueTime = Timeout.InfiniteTimeSpan;
         var period = Timeout.InfiniteTimeSpan;
-        using var sut = new TestScheduler();
+        var sut = new TestScheduler();
         using var timer = sut.CreateTimer(_ => callbackCount++, null, dueTime, period);
 
         sut.ForwardTime(TimeSpan.FromSeconds(1));
@@ -56,7 +56,7 @@ public class TestSchedulerTimerTests
     {
         // Arrange
         var callbackCount = 0;
-        using var sut = new TestScheduler();
+        var sut = new TestScheduler();
         using var timer = sut.CreateTimer(_ => callbackCount++, null, Timeout.InfiniteTimeSpan, Timeout.InfiniteTimeSpan);
         var dueTime = TimeSpan.FromSeconds(1);
         var period = TimeSpan.FromSeconds(2);
@@ -80,7 +80,7 @@ public class TestSchedulerTimerTests
     {
         // Arrange
         var callbackCount = 0;
-        using var sut = new TestScheduler();
+        var sut = new TestScheduler();
         var originalDueTime = TimeSpan.FromSeconds(3);
         var period = TimeSpan.FromSeconds(5);
         using var timer = sut.CreateTimer(_ => callbackCount++, null, originalDueTime, period);
@@ -98,5 +98,36 @@ public class TestSchedulerTimerTests
 
         sut.ForwardTime(period);
         callbackCount.Should().Be(2);
+    }
+
+    [Fact]
+    public void Timer_callback_invoked_multiple_times_single_forward()
+    {
+        var callbackCount = 0;
+        var sut = new TestScheduler();
+        var dueTime = TimeSpan.FromSeconds(3);
+        var period = TimeSpan.FromSeconds(5);
+        using var timer = sut.CreateTimer(_ => callbackCount++, null, dueTime, period);
+
+        sut.ForwardTime(TimeSpan.FromSeconds(13));
+
+        callbackCount.Should().Be(3);
+    }
+
+    [Fact]
+    public void GetUtcNow_matches_time_at_callback_time()
+    {
+        var sut = new TestScheduler();
+        var startTime = sut.GetUtcNow();
+        var callbackTimes = new List<DateTimeOffset>();
+        var interval = TimeSpan.FromSeconds(3);
+        using var timer = sut.CreateTimer(_ => callbackTimes.Add(sut.GetUtcNow()), null, interval, interval);
+
+        sut.ForwardTime(interval * 3);
+
+        callbackTimes.Should().Equal(
+            startTime + interval * 1,
+            startTime + interval * 2,
+            startTime + interval * 3);
     }
 }